#109 Implemented automated testing to retrieve all route suggestions
This test checks the route generation process. It goes from the home page to the routes page. Then, it enters a set to and from destination and prompts the system for directions. Finally, it makes sure that the system has produced the correct number of suggestions per method of transport