#27 Updated GUI test
As the previous gui test just clicked on the next input bar, the test was breaking due to the "from" bar being hidden by the suggest list of POIs. To fix this, the test now selects a POI from the drop down menu.

Also, the "Get Directions" button was removed as entering a "to" and "from" location now automatically generates suggested routes